# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AuditIQ API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  ASSOCIATE: "ASSOCIATE",
  SENIOR: "SENIOR",
  MANAGER: "MANAGER",
} as const;

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
  role?: RegisterBodyRole;
  firmName?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  ASSOCIATE: "ASSOCIATE",
  SENIOR: "SENIOR",
  MANAGER: "MANAGER",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  firmName?: string;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type EngagementStatus =
  (typeof EngagementStatus)[keyof typeof EngagementStatus];

export const EngagementStatus = {
  ACTIVE: "ACTIVE",
  COMPLETED: "COMPLETED",
  ARCHIVED: "ARCHIVED",
} as const;

export interface Engagement {
  id: number;
  name: string;
  clientName: string;
  period: string;
  status: EngagementStatus;
  totalEntries?: number;
  highRiskCount?: number;
  mediumRiskCount?: number;
  lowRiskCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEngagementBody {
  name: string;
  clientName: string;
  period: string;
  description?: string;
}

export interface UploadResponse {
  message: string;
  engagementId: number;
  totalRows: number;
  processedRows: number;
  errors?: string[];
}

export type RiskScoreRiskLevel =
  (typeof RiskScoreRiskLevel)[keyof typeof RiskScoreRiskLevel];

export const RiskScoreRiskLevel = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export interface RiskScore {
  id: number;
  entryId: number;
  totalScore: number;
  riskLevel: RiskScoreRiskLevel;
  postingTimeScore?: number;
  amountScore?: number;
  userConcentrationScore?: number;
  keywordScore?: number;
  frequencyScore?: number;
  confidenceScore: number;
  overridden: boolean;
  overrideReason?: string;
  overriddenBy?: string;
  overriddenAt?: string;
  createdAt?: string;
}

export interface JournalEntry {
  id: number;
  engagementId: number;
  entryDate: string;
  postedBy: string;
  description: string;
  debitAccount?: string;
  creditAccount?: string;
  amount: number;
  postingTime?: string;
  referenceNumber?: string;
  riskScore?: RiskScore;
}

export type JournalEntryDetailRawData = { [key: string]: unknown };

export interface AiExplanation {
  id: number;
  entryId: number;
  explanation: string;
  triggers: string[];
  isaReference?: string;
  generatedAt: string;
  modelUsed?: string;
}

export interface JournalEntryDetail {
  id: number;
  engagementId: number;
  entryDate: string;
  postedBy: string;
  description: string;
  debitAccount?: string;
  creditAccount?: string;
  amount: number;
  postingTime?: string;
  referenceNumber?: string;
  rawData?: JournalEntryDetailRawData;
  riskScore?: RiskScore;
  aiExplanation?: AiExplanation;
}

export interface JournalEntryListResponse {
  entries: JournalEntry[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export type OverrideBodyRiskLevel =
  (typeof OverrideBodyRiskLevel)[keyof typeof OverrideBodyRiskLevel];

export const OverrideBodyRiskLevel = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export interface OverrideBody {
  riskLevel: OverrideBodyRiskLevel;
  reason: string;
}

export interface UserRiskSummary {
  user: string;
  entryCount: number;
  highRiskCount: number;
  avgScore: number;
  totalAmount: number;
}

export interface DashboardSummary {
  engagementId: number;
  totalEntries: number;
  highRiskCount: number;
  mediumRiskCount: number;
  lowRiskCount: number;
  highRiskPct?: number;
  mediumRiskPct?: number;
  lowRiskPct?: number;
  avgRiskScore: number;
  topRiskyUsers?: UserRiskSummary[];
  recentHighRisk?: JournalEntry[];
  afterHoursCount?: number;
  weekendCount?: number;
  aiExplanationCount?: number;
  duplicateSuspects?: number;
}

export interface UserHeatmapRow {
  user: string;
  totalEntries: number;
  highRiskEntries: number;
  avgRiskScore: number;
  totalAmount: number;
  riskConcentration: number;
}

export interface TimeHeatmapRow {
  hour: number;
  dayOfWeek: number;
  entryCount: number;
  avgRiskScore: number;
  label?: string;
}

export interface ScoreRange {
  range: string;
  count: number;
}

export interface RiskDistribution {
  high: number;
  medium: number;
  low: number;
  scoreBreakdown: ScoreRange[];
}

export interface BenfordDigit {
  digit: number;
  expected: number;
  actual: number;
  count: number;
  deviation: number;
}

export interface BenfordAnalysis {
  engagementId: number;
  digits: BenfordDigit[];
  deviationScore: number;
  riskAssessment: string;
  summary: string;
}

export interface DuplicateGroup {
  groupId: string;
  entries: JournalEntry[];
  matchReason: string;
  riskScore: number;
}

export interface OverallDashboard {
  totalEngagements: number;
  activeEngagements: number;
  totalEntriesProcessed: number;
  totalHighRisk: number;
  totalMediumRisk: number;
  totalLowRisk: number;
  recentEngagements?: Engagement[];
  overallRiskTrend?: string;
}

export interface AuditLog {
  id: number;
  engagementId?: number;
  userId?: number;
  action: string;
  entityType: string;
  entityId?: number;
  details?: string;
  ipAddress?: string;
  createdAt: string;
}

export type ListJournalEntriesParams = {
  riskLevel?: ListJournalEntriesRiskLevel;
  user?: string;
  dateFrom?: string;
  dateTo?: string;
  page?: number;
  pageSize?: number;
};

export type ListJournalEntriesRiskLevel =
  (typeof ListJournalEntriesRiskLevel)[keyof typeof ListJournalEntriesRiskLevel];

export const ListJournalEntriesRiskLevel = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export type ListAuditLogsParams = {
  engagementId?: number;
  page?: number;
};