# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AuditIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /engagements/{id}/heatmap/users:
    get:
      operationId: getUserHeatmap
      tags: [dashboard]
      summary: Get risk heatmap by user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserHeatmapRow"
  /engagements/{id}/heatmap/time:
    get:
      operationId: getTimeHeatmap
      tags: [dashboard]
      summary: Get risk heatmap by posting time
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Time heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeHeatmapRow"
  /engagements/{id}/risk-distribution:
    get:
      operationId: getRiskDistribution
      tags: [dashboard]
      summary: Get risk distribution breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskDistribution"
  /engagements/{id}/benford:
    get:
      operationId: getBenfordAnalysis
      tags: [dashboard]
      summary: Get Benford's Law analysis for an engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Benford analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenfordAnalysis"
  /engagements/{id}/duplicates:
    get:
      operationId: getDuplicates
      tags: [dashboard]
      summary: Get potential duplicate entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Duplicate groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DuplicateGroup"

  # Reports
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          type: string
          enum: [ASSOCIATE, SENIOR, MANAGER]
          default: ASSOCIATE
        firmName:
          type: string
      required: [email, password, name]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [ASSOCIATE, SENIOR, MANAGER]
        firmName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role]

    # Engagements
...
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        entryDate:
          type: string
        postedBy:
          type: string
        description:
          type: string
        debitAccount:
          type: string
        creditAccount:
          type: string
        amount:
          type: number
        postingTime:
          type: string
        referenceNumber:
          type: string
        riskScore:
          $ref: "#/components/schemas/RiskScore"
      required: [id, engagementId, entryDate, postedBy, description, amount]
...
    RiskScore:
      type: object
      properties:
        id:
          type: integer
        entryId:
          type: integer
        totalScore:
          type: number
        riskLevel:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        postingTimeScore:
          type: number
        amountScore:
          type: number
        userConcentrationScore:
          type: number
        keywordScore:
          type: number
        frequencyScore:
          type: number
        confidenceScore:
          type: number
        overridden:
          type: boolean
        overrideReason:
          type: string
        overriddenBy:
          type: string
        overriddenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, entryId, totalScore, riskLevel, confidenceScore, overridden]
...
    UserHeatmapRow:
      type: object
      properties:
        user:
          type: string
        totalEntries:
          type: integer
        highRiskEntries:
          type: integer
        avgRiskScore:
          type: number
        totalAmount:
          type: number
        riskConcentration:
          type: number
      required: [user, totalEntries, highRiskEntries, avgRiskScore, totalAmount, riskConcentration]
    TimeHeatmapRow:
      type: object
      properties:
        hour:
          type: integer
        dayOfWeek:
          type: integer
        entryCount:
          type: integer
        avgRiskScore:
          type: number
        label:
          type: string
      required: [hour, dayOfWeek, entryCount, avgRiskScore]
    RiskDistribution:
      type: object
      properties:
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
        scoreBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ScoreRange"
      required: [high, medium, low, scoreBreakdown]
    ScoreRange:
      type: object
      properties:
        range:
          type: string
        count:
          type: integer
      required: [range, count]
    BenfordAnalysis:
      type: object
      properties:
        engagementId:
          type: integer
        digits:
          type: array
          items:
            $ref: "#/components/schemas/BenfordDigit"
        deviationScore:
          type: number
        riskAssessment:
          type: string
        summary:
          type: string
      required: [engagementId, digits, deviationScore, riskAssessment, summary]
    BenfordDigit:
      type: object
      properties:
        digit:
          type: integer
        expected:
          type: number
        actual:
          type: number
        count:
          type: integer
        deviation:
          type: number
      required: [digit, expected, actual, count, deviation]
    DuplicateGroup:
      type: object
      properties:
        groupId:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/JournalEntry"
        matchReason:
          type: string
        riskScore:
          type: number
      required: [groupId, entries, matchReason, riskScore]
...